+++ tests/test_cons.py
@@ -216,11 +216,43 @@ def test_addConsDisjunction_expr_init():
     assert m.isEQ(m.getVal(y), 5)
     assert m.isEQ(m.getVal(o), 6)
 
+def test_cons_knapsack():
+    m = Model()
+    x = m.addVar("x", lb=0, ub=2, obj=-1)
+    y = m.addVar("y", lb=0, ub=4, obj=0)
+    z = m.addVar("z", lb=0, ub=5, obj=2)
+    
+    knapsack_cons = m.addCons(4*x + 2*y <= 10, knapsack=True)
 
-@pytest.mark.skip(reason="TODO: test getValsLinear()")
+    assert knapsack_cons.getConshdlrName() == "knapsack"
+    assert knapsack_cons.isKnapsack()
+
+    m.chgRhs(knapsack_cons, 5)
+
+    assert knapsack_cons.getRhs() == 5
+
+    m.addCoefKnapsack(knapsack_cons, z, 3)
+    weights = m.getWeightsKnapsack(knapsack_cons)
+    assert weights["x"] == 4
+    assert weights["y"] == 2
+    assert weights["z"] == 3
+
+    m.optimize()
+    assert m.getDualsolKnapsack(knapsack_cons) == 0
+
 def test_getValsLinear():
-    assert True
+    m = Model()
+    x = m.addVar("x", lb=0, ub=2, obj=-1)
+    y = m.addVar("y", lb=0, ub=4, obj=0)
+    z = m.addVar("z", lb=0, ub=5, obj=2)
+    
+    c1 = m.addCons(2*x + y <= 5)
+    c2 = m.addCons(x + 4*z <= 5)
+    assert m.getValsLinear(c1) == [2,1]
 
+    m.optimize() # just to check if constraint transformation matters
+
+    assert m.getValsLinear(c2) == [1,4]
 
 @pytest.mark.skip(reason="TODO: test getRowLinear()")
 def test_getRowLinear():